requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
catboost==1.2.8
category_encoders==2.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kafka-python==1.4.7
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.42.1
numpy==2.3.0
packaging==24.2
pandas==2.3.0
patsy==1.0.1
pillow==11.2.1
plotly==6.1.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
six==1.16.0
smmap==5.0.2
statsmodels==0.14.4
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
